chore: reorganize chat page to follow nextjs principles
