+++ web/personas-open-source/src/app/chat/_components/loading-fallback.tsx
@@ -1,0 +1,7 @@
+export function LoadingFallback() {
+  return (
+    <div className="flex items-center justify-center h-screen bg-black text-white">
+      <div>Loading...</div>
+    </div>
+  );
+}
